requirements.txt
absl-py==2.1.0
ale-py==0.10.1
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
Farama-Notifications==0.0.4
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.69.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
orjson==3.10.13
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pygame==2.6.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.4.1
torch-directml==0.2.5.dev240914
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
